Notes.js implementation START
Here's the thought process for the functions used in Notes.js, but do NOT work yet:

The useState hook is used for setting initial state for notes (array of existing notes) and newNote (text entered by the user in the form).

The useEffect hook in concept fetches existing notes from the database when the component mounts. The axios library is used to make the GET request.

handleNoteSubmit is called when the user submits the form, creating a new note object with the current date and the user's input, and sending a POST request to the API to save the note to the database. If the request is successful, the new note is added to the state and the input field is cleared.

handleNoteChange is called whenever the user types in the input field, updating the newNote state with the new value.

And, of course, the return renders the results in html, including a form with input field and submit button, as well as a list of existing notes fetched from the API.